to string

+++ 0x00-python_variable_annotations/3-to_str.py
@@ -3,7 +3,7 @@
 the string representation of the float. """
 
 
-def to_str(n: int) -> str:
+def to_str(n: float) -> str:
     """ This model float n as argument and returns
     the string representation of the float. """
